refactor: implement and test the BFT storage contract once
