+++ node/bft/storage-service/src/lib.rs
@@ -28,3 +28,7 @@ pub use persistent::*;
 
 pub mod traits;
 pub use traits::*;
+
+// The expectations shared by every service, which need every service to assert them against.
+#[cfg(all(test, feature = "memory", feature = "persistent"))]
+mod tests;
